fix(registry-duck-ui): done with focus this is the proper one.
