+++ packages/registry-ui-duckui/src/dropdown-menu/dropdown-menu.libs.ts
@@ -1,94 +1,51 @@
 import React from 'react'
-import { styleItem } from '../command/command.libs'
+import { dstyleItem, handleItemsSelection, styleItem } from '../command/command.libs'
 
-export function handleItemsSelection(
-  currentItem: number,
-  itemsRef: React.RefObject<HTMLLIElement[]>,
+export function initRefs(
+  groupsRef: React.RefObject<HTMLDivElement[] | null>,
+  wrapperRef: React.RefObject<HTMLDivElement | null>,
   selectedItemRef: React.RefObject<HTMLLIElement | null>,
-) {
-  // This will remove the class from all filteredItems.and add it to the right one.
-  for (let i = 0; i < itemsRef.current.length; i++) {
-    const item = itemsRef.current[i] as HTMLLIElement
-    item.blur()
-    item.removeAttribute('aria-selected')
-
-    if (i === currentItem) {
-      styleItem(item)
-      selectedItemRef.current = item
-      item.scrollIntoView({ block: 'center', behavior: 'smooth' })
-    }
-  }
-}
-
-export function useHandleKeyDown(
   itemsRef: React.RefObject<HTMLLIElement[]>,
-  selectedItemRef: React.RefObject<HTMLLIElement | null>,
   originalItemsRef: React.RefObject<HTMLLIElement[]>,
   triggerRef: React.RefObject<HTMLButtonElement | null>,
   contentRef: React.RefObject<HTMLDivElement | null>,
-  onOpenChange?: (open: boolean) => void,
 ) {
-  React.useEffect(() => {
-    const html = document.documentElement
-    let originalCurrentItem = 0
-    let currentItem = 0
-    let inSubMenu = false
+  const groups = wrapperRef.current?.querySelectorAll('[duck-dropdown-menu-group]')
+  const items = wrapperRef.current?.querySelectorAll(
+    '[duck-dropdown-menu-group]>[duck-dropdown-menu-item], [duck-dropdown-menu-group]>*>[duck-dropdown-menu-item]',
+  ) as never as HTMLLIElement[]
+  groupsRef.current = Array.from(groups ?? []) as HTMLDivElement[]
+  itemsRef.current = Array.from(items ?? []) as HTMLLIElement[]
+  originalItemsRef.current = Array.from(items ?? []) as HTMLLIElement[]
 
-    function handleKeyDown(e: KeyboardEvent) {
-      if (e.key === 'ArrowDown') {
-        const itemIndex = currentItem === itemsRef.current.length - 1 ? 0 : currentItem + 1
-        currentItem = itemIndex
-        if (!inSubMenu) originalCurrentItem = itemIndex
-      } else if (e.key === 'ArrowUp') {
-        const itemIndex = currentItem === 0 ? itemsRef.current.length - 1 : currentItem - 1
-        currentItem = itemIndex
-        if (!inSubMenu) originalCurrentItem = itemIndex
-      } else if (e.key === 'Enter') {
-        ;(itemsRef.current[currentItem] as HTMLLIElement)?.click()
-        if (onOpenChange) onOpenChange(false)
-        contentRef.current?.setAttribute('data-open', 'false')
-        triggerRef.current?.setAttribute('aria-open', 'false')
-      }
+  const selectedITem = itemsRef.current.find((item) => !item.hasAttribute('disabled'))
 
-      if (
-        (e.key === 'ArrowLeft' && html.getAttribute('dir') === 'rtl') ||
-        (e.key === 'ArrowRight' && (html.getAttribute('dir') === 'ltr' || html.getAttribute('dir') === null))
-      ) {
-        const item = itemsRef.current[originalCurrentItem] as HTMLLIElement
-        const parent = item?.parentNode as HTMLDivElement
-        if (!parent?.hasAttribute('duck-dropdown-menu-sub')) return
+  styleItem(selectedITem ?? null)
+  selectedITem?.focus()
+  itemsRef.current = itemsRef.current.filter(
+    (item) => !(item.hasAttribute('disabled') || item.getAttribute('disabled') === 'true'),
+  )
+  console.log(itemsRef.current)
+  selectedItemRef.current = selectedITem ?? null
 
-        const subItems = Array.from(parent?.querySelectorAll('[duck-dropdown-menu-item]') as never as HTMLLIElement[])
-          .splice(1, 3)
-          .filter((item) => !(item.hasAttribute('disabled') || item.getAttribute('disabled') === 'true'))
+  for (let i = 0; i < itemsRef.current?.length; i++) {
+    const item = itemsRef.current[i] as HTMLLIElement
 
-        if (subItems.length <= 0) return
+    item.addEventListener('mouseover', () => {
+      if (item.hasAttribute('aria-checked')) return
+      dstyleItem(item)
+      item?.blur()
+    })
 
-        item.setAttribute('data-open', 'true')
-        itemsRef.current = subItems
-        inSubMenu = true
-        currentItem = 0
-      }
-
-      if (
-        (e.key === 'ArrowRight' && html.getAttribute('dir') === 'rtl') ||
-        (e.key === 'ArrowLeft' && (html.getAttribute('dir') === 'ltr' || html.getAttribute('dir') === null))
-      ) {
-        const subItem = itemsRef.current[currentItem] as HTMLLIElement
-        subItem.removeAttribute('aria-selected')
-        itemsRef.current = originalItemsRef.current.filter((item) => !item.hasAttribute('disabled'))
-
-        const item = itemsRef.current[originalCurrentItem] as HTMLLIElement
-        item.setAttribute('data-open', 'false')
-
-        inSubMenu = false
-        currentItem = originalCurrentItem
-      }
-
-      handleItemsSelection(currentItem, itemsRef, selectedItemRef)
-    }
-
-    document.addEventListener('keydown', handleKeyDown)
-    return () => document.removeEventListener('keydown', handleKeyDown)
-  }, [])
+    item.addEventListener('click', () => {
+      const currentItem = itemsRef.current?.findIndex((_item) => _item.id === item.id)
+      handleItemsSelection(currentItem, itemsRef, () => {
+        selectedItemRef.current = item
+      })
+      wrapperRef.current?.querySelector('[duck-select-value]')?.setHTMLUnsafe(item.children[0]?.getHTML() ?? '')
+      item.setAttribute('aria-checked', 'true')
+      triggerRef.current?.setAttribute('data-open', 'false')
+      contentRef.current?.setAttribute('data-open', 'false')
+    })
+  }
 }
